release/4.15.0 (#71)

+++ sql/uninstall.php
@@ -27,6 +27,10 @@
 * to avoid any conflicts with others containers.
 */
 
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 //old versions
 Configuration::deleteByName('MERCADOPAGO_CLIENT_ID');
 Configuration::deleteByName('MERCADOPAGO_CLIENT_SECRET');
